+++ tests/http-trailers_test.cpp
@@ -1,6 +1,5 @@
 #include <gtest/gtest.h>
 
-#include <chrono>
 #include <string>
 #include <string_view>
 #include <utility>